ex4.15-4.22

+++ bloglist/controllers/blogs.js
@@ -1,8 +1,9 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', {username: 1, name:1})
 
   response.json(blogs)
 })
@@ -18,26 +19,38 @@ blogsRouter.get('/:id', async (request, response) => {
   
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', middleware.tokenExtractor, middleware.userExtractor, async (request, response) => {
 
   const body = request.body
+
+  const user = request.user
 
   const newBlog = new Blog ({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes || 0,
+    user: user._id
   })
 
   const savedBlog = await newBlog.save()
+  user.blogs = user.blogs.concat(savedBlog._id) 
+  await user.save()
 
   response.json(savedBlog.toJSON())
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
+blogsRouter.delete('/:id', middleware.tokenExtractor, middleware.userExtractor, async (request, response) => {
 
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+
+  if(blog.user.toString() === request.user.id.toString()){
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }else{
+    return response.status(401).json({error : 'you do not have the access to delete this blog'})
+  }
+
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
